ENH: rename server.py to isdi
It is somewhat unintuitive that server.py is the main entrypoint of the
isdi platform, yet it's not clear it is meant to be run as a webserver.
Likewise, the mentions on README.md should be clearer that, as of now,
the flask webserver is the main way to interact with the application.

As an addition, run.sh seems to be making some opinionated decisions on
how to set things up for server.py, and it includes hardcoded, absolute user
paths that do not work in the general case. As such, it is better to
remove it.