config: remove hardcoded DEBUG/TEST variables
It appears that the DEBUG/TEST checks using environment variables were
shadowed with a hardcoded 'False' for both, which made it impossible to
deploy the app in debug/test mode without manually editing the
config.py file. Remove those lines so as the environment variables DEBUG
and TEST can be used.