web.model: add web module, move client model to it
The primary isdi script should be used as an entrypoint and where things
are set up, not defined. Having all the code live in a monolithic
repository harms readability. Move the client class definition to its
own module to help with readability and code re-use.

This is part of a major refactoring of the isdi flask application, which
explains the somewhat elaborate directory structure chosen in this
commiit (web/client.py) rather than (client.py).