web:view:consult: move consult forms to module
The consultation form views can easily live in their own module. This
will help with refactoring and improving the codebase. Although at first
sight this may need to live in the forms submodule, these forms are
actually views. This is somewhat of a misnomer that will be fixed in
future commits.